backend/chalicelib_project/requirements.txt
boto3==1.26.104
chalice==1.28.0
openai==1.6.1
requests==2.28.2
slack-sdk==3.26.1
pydantic==2.5.0
cachetools==5.3.2
python-dotenv==1.0.0
# Start QA
psycopg2==2.9.9
SQLAlchemy==2.0.23
marshmallow==3.20.1
inflection==0.5.1
pgvector==0.2.4
langchain==0.0.352
tiktoken==0.5.2
sentence-transformers==2.2.2
unstructured==0.11.6
pandas<=2.1.4
werkzeug<=3.0.1
nltk<=3.8.1
langchain==0.0.352
# End QA
